i1006 Fixes and CI's for point scoring
There were a few issues with "manual" judging and point scoring contests, which also brought to light issues with auto judging.
Use correct judgment result string instead of the CLICS string.  We have to map result strings to those defined by PC2 (and reject.ini).
CI: get rid of annoying prompt telling you the problem has changed when you go to edit it, but in fact, it has not.
Update GUI for judge so it shows point scoring "score" values and, actually updates the score (it did not previously - apparently this was not done when point scoring autojudging was implemented).
Do not show score column for test case results if it's not a point scoring contest.